+++ import_export/admin.py
@@ -1,6 +1,5 @@
 import logging
 import warnings
-from contextlib import contextmanager
 
 import django
 from django import forms
@@ -68,24 +67,6 @@ class ImportExportMixinBase:
         extra_context = extra_context or {}
         extra_context["base_change_list_template"] = self.base_change_list_template
         return super().changelist_view(request, extra_context)
-
-
-class FakePaginator:
-    count = 0
-
-
-def _get_paginator(request, queryset, per_page):
-    return FakePaginator()
-
-
-@contextmanager
-def temp_attr(obj, attr_name, new_value):
-    original_value = getattr(obj, attr_name)
-    setattr(obj, attr_name, new_value)
-    try:
-        yield
-    finally:
-        setattr(obj, attr_name, original_value)
 
 
 class ImportMixin(BaseImportMixin, ImportExportMixinBase):
@@ -761,12 +742,24 @@ class ExportMixin(BaseExportMixin, ImportExportMixinBase):
         if django.VERSION >= (4, 0):
             changelist_kwargs["search_help_text"] = self.search_help_text
 
-        # Temporarily set to False to avoid unnecessary COUNT queries.
-        with temp_attr(self, "show_full_result_count", False):
-            # Temporarily set to FakePaginator to avoid unnecessary COUNT queries.
-            with temp_attr(self, "get_paginator", _get_paginator):
-                cl = ChangeList(**changelist_kwargs)
+        class ExportChangeList(ChangeList):
+            def get_results(self, request):
+                """
+                We override this method because we only call ChangeList.get_queryset()
+                so we don't need anything from this method.
+                The get_results() gets called during ChangeList.__init__()
+                and we do want to avoid unnecessary COUNT queries.
+                """
+                pass
 
+        cl = ExportChangeList(**changelist_kwargs)
+
+        # get_queryset() is already called during initialization,
+        # it is enough to get it's results
+        if hasattr(cl, "queryset"):
+            return cl.queryset
+
+        # Fallback in case the ChangeList doesn't have queryset attribute set
         return cl.get_queryset(request)
 
     def get_export_data(self, file_format, queryset, *args, **kwargs):